requirements.txt
numpy>=1.21.0
pyqtgraph>=0.13.0
PyQt6>=6.0.0
spidev>=3.5
lgpio-compat>=0.1.0
